Log input gpio binary state value instead of T/F
On gpio assertion/de-assertion events we were logging
T/F, but we are switching to printing their binary values
0/1 to enhance readability

Fixes jira https://

+++ src/platform/nvl144/nvl144_power_control.cpp
@@ -74,7 +74,8 @@ NVL144PowerControl::NVL144PowerControl(
 // NVL144-specific GPIO handler implementations
 void NVL144PowerControl::nvl144pdbMainPowerOkHandler(bool state)
 {
-    lg2::info("NVL144PDBMainPowerOk GPIO event: value={VALUE}", "VALUE", state);
+    lg2::info("NVL144PDBMainPowerOk GPIO event: value={VALUE}", "VALUE",
+              static_cast<int>(state));
 
     auto it = powerSignalMap.find("NVL144PDBMainPowerOk");
     if (it == powerSignalMap.end())
